DarkMatterProfiles

Documentation for DarkMatterProfiles.jl.

Interfaces

DarkMatterProfiles.dmdensity — Method

dmdensity(p::DMProfile, r)
dmdensity(p::DMProfile, x, y, z)

Return dark matter density of the profile p in Galactocentric coordinates.

Example

julia> dmp = DMPEinasto(0.3, 8.5, 20.0, 0.17)
DMPEinasto{Float64, Float64, Float64}(0.3, 8.5, 20.0, 0.17)

julia> dmdensity(dmp, 8.5)
0.3

DarkMatterProfiles.dmdensity_galactic — Method

dmdensity_galactic(p::DMProfile, r, b, l; rsun=8.5)

Return dark matter density of the profile p in Galactic coordinates.

Arguments

p::DMProfile: The dark matter profile.

r: The distance along the line-of-sight.

b: The Galactic latitude.

l: The Galactic longitude.

Keywords

rsun: The distance from the Sun to the Galactic Center.

DarkMatterProfiles.DMProfile — Method

(p::DMProfile)(r)
(p::DMProfile)(x, y, z)

DMProfiles as functors.

Example

julia> dmp = DMPEinasto()
DMPEinasto{Float64, Float64, Float64}(0.3, 8.5, 20.0, 0.17)

julia> dmp(8.5)
0.3

Dark Matter Profiles

DarkMatterProfiles.DMPEinasto — Type

Einasto Profile

\[ ρ^\text{EIN}_χ(r) = ρ_χ^\text{loc} \exp\left[ -\frac{2}{α}\left(\frac{r^α-R_{⊙}^α} {R_\text{s}^α}\right) \right]\]

Fields

rho0::T<:Number=0.3: $ρ_χ^\text{loc}$

rsun::U<:Number=8.5: $R_⊙$

rs::U<:Number=20.0: $R_\text{s}$

alpha::V<:Number=0.17: $α$

DarkMatterProfiles.DMPNFW — Type

NFW Profile (generalized)

\[ ρ^\text{NFW}_χ(r) = ρ_χ^\text{loc} \left(1 + \frac{R_⊙}{R_\text{s}}\right)^{3-γ} \left(\frac{r}{R_⊙}\right)^{-γ} \left(1 + \frac{r}{R_\text{s}}\right)^{γ-3}\]

for $r > R_\text{cut}$, otherwise $ρ^\text{NFW}_χ(r) = ρ^\text{NFW}_χ(R_\text{cut})$. The default value is $R_\text{cut} = 0$. Note that the Schwarzschild radius of the supermassive black hole at the Galactic Center is $R = 2GM/c^2 ≈ 1.2 × 10^7~\text{km} ≈ 4 × 10^{-10}~\text{kpc}$, where $M ≈ 4.15 × 10^6 M_⊙$.

Fields

rho0::T<:Number=0.3: $ρ_χ^\text{loc}$

rsun::U<:Number=8.5: $R_⊙$

rs::U<:Number=20.0: $R_\text{s}$

rcut::U<:Number=0: $R_\text{cut}$

gamma::V<:Number=1: $γ$

Index

DarkMatterProfiles.DMPEinasto

DarkMatterProfiles.DMPNFW

DarkMatterProfiles.DMProfile

DarkMatterProfiles.dmdensity

DarkMatterProfiles.dmdensity_galactic