DarkMatterProfiles

Documentation for DarkMatterProfiles.

Interfaces

DarkMatterProfiles.dmdensity — Method

dmdensity(r::Real, p::DMProfile)
dmdensity(x::Real, y::Real, z::Real, p::DMProfile)

Return dark matter density of the profile p in Galactocentric coordinates.

Example

dmp = EinastoProfile(0.3, 8.5, 20.0, 0.17)
dmdensity(8.5, dmp)

DarkMatterProfiles.dmdensity_galactic — Method

dmdensity_galactic(r::Real, b::Real, l::Real, p::DMProfile; rsun::Real=8.5)

Return dark matter density of the profile p in Galactic coordinates.

Arguments

r::Real: The distance along the line-of-sight.

b::Real: The Galactic latitude.

l::Real: The Galactic longitude.

p::DMProfile: The dark matter profile.

Keywords

rsun::Real: The distance from the Sun to the Galactic Center.

DarkMatterProfiles.DMProfile — Method

DMProfile as functor

Dark Matter Profiles

DarkMatterProfiles.EinastoProfile — Type

Einasto Profile

\[\rho^\mathrm{EIN}_\chi(r) = \rho_\chi^\text{loc} \exp\left[-\frac{2}{\alpha}\left(\frac{r^\alpha-R_{\odot}^\alpha} {R_\text{s}^\alpha}\right)\right]\]

Fields

rho0::Real: $\rho_\chi^\text{loc}$

rsun::Real: $R_\odot$

rs::Real: $R_\text{s}$

alpha::Real: $\alpha$

Index

DarkMatterProfiles.DMProfile

DarkMatterProfiles.EinastoProfile

DarkMatterProfiles.dmdensity

DarkMatterProfiles.dmdensity_galactic